# technology-solutions-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        ":80/api/perfis": {
            "get": {
                "summary": "Get all profiles",
                "operationId": "c6ad2e48ee09c980b83b1dc6bd824ad7",
                "responses": {
                    "200": {
                        "description": "Profiles retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Perfis retornados com sucesso"
                                        },
                                        "object": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Perfil"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No profiles found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Nenhum perfil encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        ":80/api/perfis/criar": {
            "post": {
                "summary": "Create a new profile",
                "operationId": "8749a149237502f55ce5c90e4e4e099c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "perfil",
                                    "status"
                                ],
                                "properties": {
                                    "perfil": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Perfil cadastrado com sucesso"
                                        },
                                        "object": {
                                            "$ref": "#/components/schemas/Perfil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Erro de validação."
                                        },
                                        "object": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        ":80/api/perfis/editar/{cpf}/{id_perfil}": {
            "patch": {
                "summary": "Update a profile",
                "operationId": "00a1b3998e9603aee241a0768f35069a",
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "CPF of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id_perfil",
                        "in": "path",
                        "description": "ID of the profile",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Perfil atualizado com sucesso"
                                        },
                                        "object": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Erro de validação."
                                        },
                                        "object": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Perfil não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Perfil": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "perfil": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "status": {
                        "type": "string",
                        "example": "A"
                    },
                    "data": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    }
                },
                "type": "object"
            }
        }
    }
}